SeriesContext: Move to own file
It starts to gather logic, so better move it into its own file.
