+++ src/config/mod.rs
@@ -1,4 +1,5 @@
 pub mod meta_data;
 pub mod phenoxtracter_config;
 pub mod pipeline_config;
+pub mod series_context;
 pub mod table_context;
