+++ src/lib.rs
@@ -1,9 +1,7 @@
 mod config;
+mod error;
 mod extract;
 mod load;
 mod pipeline;
 mod transform;
-
 mod validation;
-
-mod error;
